+++ foxtive-ntex/src/lib.rs
@@ -7,5 +7,5 @@ pub mod http;
 pub mod setup;
 
 pub use setup::state::AppState;
-pub use http::shutdown::{ShutdownConfig, ShutdownRegistry};
+pub use http::shutdown::{ShutdownConfig, ShutdownRegistry, ShutdownSignal};
 pub use http::server::ServerBuilder;