+++ foxtive-ntex/src/setup/mod.rs
@@ -6,7 +6,7 @@ use std::collections::HashMap;
 pub mod state;
 
 /// Custom state builder function type
-type CustomStateBuilder = Box<dyn FnOnce() -> HashMap<String, Box<dyn Any + Send + Sync>> + Send>;
+pub(crate) type CustomStateBuilder = Box<dyn FnOnce() -> HashMap<String, Box<dyn Any + Send + Sync>> + Send>;
 
 /// Internal structure used during server bootstrapping
 pub struct NtexSetup {